% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateResultMatrix.R
\name{CreateResultMatrix}
\alias{CreateResultMatrix}
\title{Creating a Result Table from a SIModel object}
\usage{
CreateResultMatrix(modelList)
}
\arguments{
\item{modelList}{A list of SIModel inputs and outputs}
}
\value{
A result table
}
\description{
Creating a Result Table from a SIModel object
}
\keyword{internal}
