% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SIModel}
\alias{print.SIModel}
\title{Title}
\usage{
\method{print}{SIModel}(out, digits = 2)
}
\arguments{
\item{out}{A SIModel object}

\item{digits}{Number of digits to round to}
}
\description{
Title
}
\keyword{internal}
