% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{result.table}
\alias{result.table}
\title{Creating a Result Table from a SIModel object}
\usage{
result.table(modelList)
}
\arguments{
\item{modelList}{A list of SIModel inputs and outputs}
}
\value{
A result table
}
\description{
Creating a Result Table from a SIModel object
}
\keyword{internal}
