% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SymbMultiplication.R
\name{\%*\%}
\alias{\%*\%}
\alias{SymbMultiplication}
\title{Internal function used to do symbolic matrix multiplication using the Ryacas
R package}
\usage{
SymbMultiplication(x1, x2)
}
\arguments{
\item{x1}{A Ryacas object.}

\item{x2}{A Ryacas object.}
}
\description{
Internal function used to do symbolic matrix multiplication using the Ryacas
R package
}
\examples{
blueprint <- matrix( c("ARx", "CLxy", "ARy", "0"), nrow = 2, ncol = 2 )

SymbolicCovMat <- matrix( c("VarX", "CovXY", "CovXY", "VarY"), ncol = 2, nrow = 2)

B <- Ryacas::ysym(blueprint)
Cov1 <- Ryacas::ysym(SymbolicCovMat)

SymbMultiplication(x1 = t(B), x2= Cov1)
}
