% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSymbCovMatrix.R
\name{GetSymbCovMatrix}
\alias{GetSymbCovMatrix}
\title{Create a symbolic covariance matrix}
\usage{
GetSymbCovMatrix(blueprint)
}
\arguments{
\item{blueprint}{A character matrix which specifies which effects to estimate and
which effects to constrain to a non-zero value}
}
\value{
A character matrix
}
\description{
Create a symbolic covariance matrix
}
\examples{
blueprint <- matrix( c("ARx", "CLxy", "ARy", "0"), nrow = 2, ncol = 2 )
colnames(blueprint) <- c("X", "Y")
rownames(blueprint) <- c("X_0", "Y_0")
GetSymbCovMatrix(blueprint)
}
