% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM.R
\name{SIM}
\alias{SIM}
\title{Estimate a Stability Informed Model}
\usage{
SIM(data = NULL, S = NULL, n = NULL, model, stability)
}
\arguments{
\item{data}{A dataframe with the measured variables. Not needed if S is provided}

\item{S}{A covariance matrix for the measured variables. Not needed if data is provided.}

\item{n}{Number of observations. Not needed if data is provided.}

\item{model}{An object with the cross-sectional model description in lavaan syntax}

\item{stability}{A data frame that contains stability information for each
variable in the model. If unnamed SIM will assume the stability
values are in the same order as the provided data set/
covariance matrix.}
}
\value{
An object of class SIM
}
\description{
Estimate a Stability Informed Model
}
\examples{

model <- 'Y~X'
stability <- data.frame(X = .3, Y = .3)
dat <- data.frame(Y = rnorm(500, 0, 1), X = rnorm(500, 0, 1), Z = rnorm(500, 0, 1))

SIM(data = dat, model = model, stability = stability)
}
